slime-inspect got a :mode argument that controls its behavior.
Most notably with :mode :dwim (default) it'll try to find out the user's
intention (e.g. look up definitions when a symbol is inspected). To avoid
annoyance at misses, it puts the original object on the inspector stack,
so that 'l' brings back to the non-dwim value.

Also add an *inspector-lookup-hooks* customization point.